+++ framework/domain/src/test/java/is/codion/framework/domain/entity/DefaultEntityTest.java
@@ -117,7 +117,7 @@ public class DefaultEntityTest {
 		assertEquals(1, fromFile.size());
 		Entity entityFromFile = (Entity) fromFile.get(0);
 		assertEquals(Detail.TYPE, entity.entityType());
-		assertTrue(entity.valuesEqual(entityFromFile));
+		assertTrue(entity.equalValues(entityFromFile));
 		assertTrue(entityFromFile.modified());
 		assertTrue(entityFromFile.modified(Detail.STRING));
 		assertEquals(originalStringValue, entityFromFile.original(Detail.STRING));
@@ -157,7 +157,7 @@ public class DefaultEntityTest {
 						detailString, detailDate, detailTimestamp, detailBoolean, referencedEntityValue);
 		test.set(testEntity);
 		assertEquals(test, testEntity, "Entities should be equal after a calling set()");
-		assertTrue(test.valuesEqual(testEntity), "Entity values should be equal after a calling set()");
+		assertTrue(test.equalValues(testEntity), "Entity values should be equal after a calling set()");
 
 		assertTrue(test.set(test).isEmpty());
 
@@ -195,20 +195,20 @@ public class DefaultEntityTest {
 						.build();
 
 		assertEquals(0, original.set(entity).size());
-		assertTrue(Entity.valuesEqual(original, entity));
+		assertTrue(original.equalValues(entity));
 
 		original.put(Detail.BOOLEAN, true);
 		entity.put(Detail.BOOLEAN, false);
 
 		assertEquals(1, original.set(entity).size());
-		assertTrue(Entity.valuesEqual(original, entity));
+		assertTrue(original.equalValues(entity));
 
 		original.put(Detail.INT, 1);
 		entity.put(Detail.INT, 2);
 		entity.put(Detail.INT, 3);//modified
 
 		assertEquals(2, original.set(entity).size());//int + int_derived
-		assertTrue(Entity.valuesEqual(original, entity));
+		assertTrue(original.equalValues(entity));
 		assertTrue(original.modified());
 		assertTrue(entity.modified());
 
@@ -218,13 +218,13 @@ public class DefaultEntityTest {
 		entity.put(Detail.STRING, "strng");
 
 		assertEquals(2, original.set(entity).size());
-		assertTrue(Entity.valuesEqual(original, entity));
+		assertTrue(original.equalValues(entity));
 
 		assertEquals(0, original.set(entity).size());
 		assertEquals(0, entity.set(original).size());
 
 		entity.remove(Detail.STRING);
-		assertFalse(Entity.valuesEqual(entity, original));
+		assertFalse(entity.equalValues(original));
 	}
 
 	@Test
@@ -414,7 +414,7 @@ public class DefaultEntityTest {
 		Entity test2 = testEntity.deepCopy();
 		assertNotSame(test2, testEntity, "Entity copy should not be == the original");
 		assertEquals(test2, testEntity, "Entities should be equal after .getCopy()");
-		assertTrue(test2.valuesEqual(testEntity), "Entity attribute values should be equal after deepCopy()");
+		assertTrue(test2.equalValues(testEntity), "Entity attribute values should be equal after deepCopy()");
 		assertNotSame(testEntity.entity(Detail.MASTER_FK), test2.entity(Detail.MASTER_FK), "This should be a deep copy");
 
 		test2.put(Detail.DOUBLE, 2.1);
@@ -468,8 +468,8 @@ public class DefaultEntityTest {
 						.with(Department.LOCATION, "Location")
 						.with(Department.ACTIVE, true)
 						.build();
-		assertTrue(dept.valuesEqual(dept.copyBuilder().build()));
-		assertFalse(dept.valuesEqual(dept.copyBuilder().with(Department.NAME, "new name").build()));
+		assertTrue(dept.equalValues(dept.copyBuilder().build()));
+		assertFalse(dept.equalValues(dept.copyBuilder().with(Department.NAME, "new name").build()));
 
 		dept.put(Department.NAME, "New name");
 		assertTrue(dept.copyBuilder().build().modified());
@@ -578,37 +578,37 @@ public class DefaultEntityTest {
 	}
 
 	@Test
-	void valuesEqual() {
+	void equalValues() {
 		Entity testEntityOne = detailEntity(detailId, detailInt, detailDouble,
 						detailString, detailDate, detailTimestamp, detailBoolean, null);
 		Entity testEntityTwo = detailEntity(detailId, detailInt, detailDouble,
 						detailString, detailDate, detailTimestamp, detailBoolean, null);
 
-		assertTrue(testEntityOne.valuesEqual(testEntityTwo));
+		assertTrue(testEntityOne.equalValues(testEntityTwo));
 
 		testEntityTwo.put(Detail.INT, 42);
-		assertFalse(testEntityOne.valuesEqual(testEntityTwo));
+		assertFalse(testEntityOne.equalValues(testEntityTwo));
 
 		testEntityOne.put(Detail.INT, 42);
-		assertTrue(testEntityOne.valuesEqual(testEntityTwo));
+		assertTrue(testEntityOne.equalValues(testEntityTwo));
 
 		testEntityTwo.remove(Detail.INT);
-		assertFalse(testEntityOne.valuesEqual(testEntityTwo));
+		assertFalse(testEntityOne.equalValues(testEntityTwo));
 
 		testEntityOne.remove(Detail.INT);
-		assertTrue(testEntityOne.valuesEqual(testEntityTwo));
+		assertTrue(testEntityOne.equalValues(testEntityTwo));
 
 		Random random = new Random();
 		byte[] bytes = new byte[1024];
 		random.nextBytes(bytes);
 
 		testEntityOne.put(Detail.BYTES, bytes);
-		assertFalse(testEntityOne.valuesEqual(testEntityTwo));
+		assertFalse(testEntityOne.equalValues(testEntityTwo));
 
 		testEntityTwo.put(Detail.BYTES, bytes);
-		assertTrue(testEntityOne.valuesEqual(testEntityTwo));
+		assertTrue(testEntityOne.equalValues(testEntityTwo));
 
-		assertThrows(IllegalArgumentException.class, () -> testEntityOne.valuesEqual(ENTITIES.entity(Master.TYPE)));
+		assertThrows(IllegalArgumentException.class, () -> testEntityOne.equalValues(ENTITIES.entity(Master.TYPE)));
 	}
 
 	@Test
@@ -774,13 +774,13 @@ public class DefaultEntityTest {
 
 		Entity copy = emp.deepCopy();
 		assertNotSame(emp, copy);
-		assertTrue(emp.valuesEqual(copy));
+		assertTrue(emp.equalValues(copy));
 		assertNotSame(emp.get(Employee.MANAGER_FK), copy.get(Employee.MANAGER_FK));
-		assertTrue(emp.entity(Employee.MANAGER_FK).valuesEqual(copy.entity(Employee.MANAGER_FK)));
+		assertTrue(emp.entity(Employee.MANAGER_FK).equalValues(copy.entity(Employee.MANAGER_FK)));
 		assertNotSame(emp.entity(Employee.MANAGER_FK).entity(Employee.DEPARTMENT_FK),
 						copy.entity(Employee.MANAGER_FK).entity(Employee.DEPARTMENT_FK));
 		assertTrue(emp.entity(Employee.MANAGER_FK).entity(Employee.DEPARTMENT_FK)
-						.valuesEqual(copy.entity(Employee.MANAGER_FK).entity(Employee.DEPARTMENT_FK)));
+						.equalValues(copy.entity(Employee.MANAGER_FK).entity(Employee.DEPARTMENT_FK)));
 
 		emp.save();
 		emp.definition().foreignKeys().get().forEach(foreignKey -> assertNotNull(emp.get(foreignKey)));