+++ changelog.md
@@ -44,6 +44,7 @@ Codion Change Log
 - FilteredTableModel.ColumnValueProvider renamed ColumnValues, related renaming.
 ### is.codion.framework.domain
 - DefaultDomain renamed DomainModel.
+- Entity.valuesEqual() instance methods renamed equalValues(), static ones removed.
 ### is.codion.framework.db.local
 - DefaultLocalEntityConnection.report() exception handling improved, now throws DatabaseException in case of an SQLException.
 - DefaultLocalEntityConnection exception handling refactored in order to guarantee rollbacks in case of RuntimeExceptions.